Streamlined the existing debugging facilities.
Everything can now be enabled/disabled from the "Debug" menu while the app is running without having to resort to the direct manipulation of defaults keys.

The "Debug" menu itself can be activated while the app is running. It will be displayed if the account is marked as a "debugger" as per the `sapo:debug` protocol or if the key combination `ctrl-option-shift` is pressed while logging in to the server. The old "IncludeDebugMenu" defaults key still works and enables the Debug menu right from the moment the app is started.

Closes #15.

git-svn-id: svn://svn.softwarelivre.sapo.pt/sapo_msg_mac/trunk@17 30d2cb3b-9f40-45d5-ba16-273292637043