Added optimizations to the CoreData store of the message center. Unread message counting is now done in constant time for each notifications channel.
In order to implement the optimization, the object model schema had to be slightly modified. So, to support the schema modification, a new schema migration mechanism was also added to the message center. Whenever the schema is changed in a future version of the app, it will be easier to implement the upgrade process for the existing data store.

Also added a new LPInternalDataUpgradeManager class that takes care of upgrading all internal data when a new version of the app is first run. It only upgrades the message center data store to the new schema for now. If the operation takes more than 0.5s, a small window with a progress bar is displayed, iTunes/iPhoto Library upgrade style.

git-svn-id: svn://svn.softwarelivre.sapo.pt/sapo_msg_mac/trunk@64 30d2cb3b-9f40-45d5-ba16-273292637043