Initial support for multi-accounts.
IMPORTANT: The implementation is partial and extremely buggy! It loves to throw exceptions around. If you don't like to live on the bleeding edge, refrain from configuring more than one account and everything should behave as expected. If it doesn't, please report the issues.

Now that the disclaimer is out of our way: Try to configure an additional account and then quit and reopen the app. You should now be connected to all the accounts that you have configured in the preferences. However, keep in mind that the circus of exception throwing should be starting shortly after.

References #60.

git-svn-id: svn://svn.softwarelivre.sapo.pt/sapo_msg_mac/trunk@90 30d2cb3b-9f40-45d5-ba16-273292637043