Some of the higher level initialization code was being run while the app was still being launched. Moved that code to a slightly later stage in the application startup process in order to allow the NSApplication exception handler to be setup appropriately before our code gets run. It also makes the app icon bounce even less in the Dock.
We were trying to report any exceptions that were raised and then ignored by the AppKit, which prevented them from reaching our top-level exception handler. However, this proved to be bad, as we were also catching some harmless exceptions that were expected in the normal flow of the application but which caused us to terminate the app. We no longer try to report back this kind of exceptions. If Apple thinks ignoring exceptions is good and doesn't let us customize that behavior, then let it be.

References #210.

git-svn-id: svn://svn.softwarelivre.sapo.pt/sapo_msg_mac/trunk@286 30d2cb3b-9f40-45d5-ba16-273292637043